Refactor filehost controllers
