+++ backend/apps/filehost/services/__init__.py
@@ -1,0 +1,1 @@
+"""Services for filehost app."""
